Refactor undergraduate programs data handling: migrate to TypeScript, enhance caching with typing support, and update YAML parsing logic.
