+++ package.json
@@ -34,6 +34,7 @@
     "@pantheon-systems/nextjs-cache-handler": "^0.7.0",
     "@plaiceholder/next": "^3.0.0",
     "@types/jsonwebtoken": "^9.0.10",
+    "@types/object-hash": "^3.0.6",
     "@uoguelph/react-components": "^1.10.1",
     "@uoguelph/web-components": "^2.1.2",
     "conditional-wrap": "^1.0.3",
